requirements.txt
annotated-types==0.7.0
anyio==4.6.2.post1
cffi==1.17.1
click==8.1.7
fastapi==0.115.3
future==1.0.0
h11==0.14.0
idna==3.10
Jinja2==3.1.4
MarkupSafe==3.0.2
numpy==2.1.3
opencv-python-headless==4.10.0.84
pillow==11.0.0
pillow_heif==0.20.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pyheif==0.8.0
python-dotenv==1.0.1
python-multipart==0.0.17
sniffio==1.3.1
starlette==0.41.0
typing_extensions==4.12.2
uvicorn==0.32.0
